// FrontFredy/src/app/model/skychain.ts
export class Skychain {
  id: Number;
  displayName: String;
  awb: Number;
  flightDate: Date;
  airlineId: Number;
  flightNumber: Number;
  piece: Number;
  weight: Number;
  cityOriginId: Number;
  cityDestinationId: Number;
  senderName: String;
  senderAddress: String;
  consigneeName: String;
  consigneeAddress: String;
  natureGoods: String;
  active: Boolean;
  userSkychain: String;
  userId: Number;
  stationId: Number;
  createdDate: Date;
  updatedDate: Date;
}